Split filesystem ext to avoid errors
